+++ target_platforms/website.py
@@ -8,6 +8,9 @@ import shutil
 import sys
 from colorama import Fore, Style
 import click
+import requests
+
+
 
 def ensure_wasm_exec(dest_dir: str = ".") -> bool:
     """
@@ -1233,48 +1236,48 @@ setup(
             files = os.listdir(os.getcwd())
             for file_name in files:
                 full_file_name = os.path.join(os.getcwd(), file_name)
-                if os.path.isfile(full_file_name):
+                if os.path.isfile(full_file_name) and file_name != 'python.7z':
                     shutil.copy(full_file_name, f"dist/{NAME}_{VERSION}")
-                elif os.path.isdir(full_file_name) and file_name != 'dist' and file_name != 'venv' and file_name != 'virtualenv' and file_name != 'node_modules':
+                elif os.path.isdir(full_file_name) and file_name != NAME and file_name != 'dist' and file_name != 'venv' and file_name != 'virtualenv' and file_name != 'node_modules':
                     shutil.copytree(full_file_name, f"dist/{NAME}_{VERSION}/{file_name}", dirs_exist_ok=True)
                 print('Copied '+file_name+' to '+f"dist/{NAME}_{VERSION}/{file_name}"+'...')
-            if not os.path.exists(f'dist/{NAME}_{VERSION}/{NAME}_app/static/logo'):
+            if not os.path.exists(f'dist/{NAME}_{VERSION}/static/logo'):
                 print('Creating logo directory...')
                 logo_directory = os.path.join(gupy_file_path, 'gupy_logo.png')       
-                os.makedirs(f'dist/{NAME}_{VERSION}/{NAME}_app/static', exist_ok=True)
-                os.makedirs(f'dist/{NAME}_{VERSION}/{NAME}_app/static/logo', exist_ok=True)
-                shutil.copy(logo_directory, f'dist/{NAME}_{VERSION}/{NAME}_app/static/logo/gupy_logo.png')
-            if not os.path.exists(f'dist/{NAME}_{VERSION}/{NAME}_app/static/splashscreen'):
+                print(logo_directory)
+                os.makedirs(f'dist/{NAME}_{VERSION}/static', exist_ok=True)
+                os.makedirs(f'dist/{NAME}_{VERSION}/static/logo', exist_ok=True)
+                shutil.copy(logo_directory, f'dist/{NAME}_{VERSION}/static/logo/gupy_logo.png')
+            if not os.path.exists(f'dist/{NAME}_{VERSION}/static/splashscreen'):
                 print('Creating splashscreen directory...')
                 splashscreen_directory = os.path.join(gupy_file_path, 'gupy_splashscreen.png')       
-                os.makedirs(f'dist/{NAME}_{VERSION}/{NAME}_app/static', exist_ok=True)
-                os.makedirs(f'dist/{NAME}_{VERSION}/{NAME}_app/static/splashscreen', exist_ok=True)
-                shutil.copy(splashscreen_directory, f'dist/{NAME}_{VERSION}/{NAME}_app/static/splashscreen/gupy_splashscreen.png')
-            if not os.path.exists(f'dist/{NAME}_{VERSION}/{NAME}_app/static/icon'):
+                os.makedirs(f'dist/{NAME}_{VERSION}/static', exist_ok=True)
+                os.makedirs(f'dist/{NAME}_{VERSION}/static/splashscreen', exist_ok=True)
+                shutil.copy(splashscreen_directory, f'dist/{NAME}_{VERSION}/static/splashscreen/gupy_splashscreen.png')
+            if not os.path.exists(f'dist/{NAME}_{VERSION}/static/icon'):
                 print('Creating icon directory...')
                 ico_directory = os.path.join(gupy_file_path, 'gupy.ico')       
-                os.makedirs(f'dist/{NAME}_{VERSION}/{NAME}_app/static', exist_ok=True)
-                os.makedirs(f'dist/{NAME}_{VERSION}/{NAME}_app/static/icon', exist_ok=True)
-                shutil.copy(ico_directory, f'dist/{NAME}_{VERSION}/{NAME}_app/static/icon/gupy.ico')
-            # package latest python if not selected - make python folder with windows/mac/linux
-            os.makedirs(f"dist/{NAME}_{VERSION}/python", exist_ok=True)
-            print('Copying python folder...')
+                os.makedirs(f'dist/{NAME}_{VERSION}/static', exist_ok=True)
+                os.makedirs(f'dist/{NAME}_{VERSION}/static/icon', exist_ok=True)
+                shutil.copy(ico_directory, f'dist/{NAME}_{VERSION}/static/icon/gupy.ico')
+                # package latest python if not selected - make python folder with windows/mac/linux
+                os.makedirs(f"dist/{NAME}_{VERSION}/python", exist_ok=True)
+                os.makedirs(f"dist/{NAME}_{VERSION}/python/macos", exist_ok=True)
+                print('Adding python dependencies...')
 
-            # import gupy_framework_windows_deps 
-            # import gupy_framework_linux_deps
-            # import gupy_framework_macos_deps
-            # gupy_framework_windows_deps.add_deps(f"dist/{NAME}_{VERSION}/python")
-            # gupy_framework_linux_deps.add_deps(f"dist/{NAME}_{VERSION}/python")
-            # gupy_framework_macos_deps.add_deps(f"dist/{NAME}_{VERSION}/python/macos")
-            # mac_pkg_file = gupy_framework_macos_deps.get_deps()[0]
-            import py7zr
-            archive_path = gupy_file_path + delim + 'python.7z'
-            with py7zr.SevenZipFile(archive_path, mode='r') as archive:
-                archive.extractall(path=f"dist/{NAME}_{VERSION}")
-            # shutil.copytree(python_loc, f"dist/{NAME}_{VERSION}/python", dirs_exist_ok=True)
-            
-            print('Copied python folder...')
-            os.chdir(f'dist/{NAME}_{VERSION}')
+                import gupy_framework_windows_deps 
+                import gupy_framework_linux_deps
+                # import gupy_framework_macos_deps
+                gupy_framework_windows_deps.add_deps(f"dist/{NAME}_{VERSION}/python")
+
+                gupy_framework_linux_deps.add_deps(f"dist/{NAME}_{VERSION}/python")
+
+                # gupy_framework_macos_deps.add_deps(f"dist/{NAME}_{VERSION}/python/macos")
+                # mac_pkg_file = gupy_framework_macos_deps.get_deps()[0]
+                    
+                
+                print('Python dependencies added.')
+                os.chdir(f'dist/{NAME}_{VERSION}')
 
 
             # command = f".{delim}python{delim}{python_folder}{delim}{python_executable} python{delim}{python_folder}{delim}get-pip.py"
